change package name, move tests
