+++ setup.py
@@ -6,7 +6,7 @@ setup(
     author="Cameron Summers, Ed Nykaza",
     author_email="cameron@tidepool.org",
     packages=['tidepool_data_science_models', 'tidepool_data_science_models.models'],
-    package_dir={'tidepool_data_science_models': 'src'},
+    package_dir={'tidepool_data_science_models': 'data_science_models'},
     license="BSD-2",
     long_description=open('README.md').read(),
     python_requires='>=3.6',
